+++ Connectors/Autocad/Speckle.Connectors.Civil3dShared/HostApp/PropertyUpdater.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.Logging;
+using Speckle.Connectors.DUI.Bindings;
 using Speckle.Sdk;
 using AAEC = Autodesk.Aec;
 using AAECPDB = Autodesk.Aec.PropertyData.DatabaseServices;
@@ -155,20 +156,4 @@ public class PropertyUpdater
       _ => value.ToString(),
     };
   }
-}
-
-public readonly struct UpdateResult
-{
-  public bool IsSuccess { get; }
-  public string? ErrorMessage { get; }
-
-  private UpdateResult(bool success, string? error)
-  {
-    IsSuccess = success;
-    ErrorMessage = error;
-  }
-
-  public static UpdateResult Success() => new(true, null);
-
-  public static UpdateResult Fail(string message) => new(false, message);
 }
