--- Connectors/Revit/Speckle.Connectors.RevitShared/HostApp/ParameterChangeRequest.cs
@@ -1,9 +1,0 @@
-namespace Speckle.Connectors.Revit.HostApp;
-
-public class ParameterChangeRequest
-{
-  public required string ApplicationId { get; init; }
-  public required string Path { get; init; }
-  public object? To { get; init; }
-  public string? InternalDefinitionName { get; set; }
-}
